+++ R/check-annotation-values.R
@@ -28,7 +28,6 @@
 #' dat2 <- data.frame(assay = "rnaSeq")
 #' check_annotation_values(dat1, annots)
 #' check_annotation_values(dat2, annots)
-#'
 #' \dontrun{
 #' syn <- synapse$Synapse()
 #' syn$login()
@@ -253,8 +252,7 @@ check_type <- function(values, key, annotations, whitelist_values = NULL,
     )
   }
 
-  correct_class <- switch(
-    coltype,
+  correct_class <- switch(coltype,
     "STRING" = "character",
     "string" = "character",
     "BOOLEAN" = "logical",
@@ -396,6 +394,13 @@ check_value <- function(values, key, annotations, whitelist_keys = NULL,
   if (all(is.na(annot_values))) {
     return(check_type(values, key, annotations, whitelist_values, return_valid))
   }
+  ## Multiple annotations come through as either comma separated list or
+  ## json-style strings (e.g. "[\"foo\",\"bar\"]"). Separate out the values.
+  if (any(grepl(",", values))) {
+    values <- gsub("\\[|\\]|\"", "", values)
+    values <- unlist(strsplit(values, split = ","))
+    values <- stringr::str_trim(values, side = "both")
+  }
   ## Check values against enumerated values in annotation definitions.
   if (isTRUE(return_valid)) {
     unique(values[values %in% c(annot_values, whitelist) & !is.na(values)])
@@ -435,6 +440,14 @@ check_value <- function(values, key, annotations, whitelist_keys = NULL,
 #' )
 #' dat <- data.frame(
 #'   fileFormat = c("wrong", "txt", "csv", "wrong again"),
+#'   stringsAsFactors = FALSE
+#' )
+#' check_values(dat, annots)
+#'
+#' ## Comma-separated and json-style strings for keys with enumerated values
+#' ## are separated and checked individually
+#' dat <- data.frame(
+#'   fileFormat = c("wrong, txt, csv", "[\"wrong again\",\"csv\"]"),
 #'   stringsAsFactors = FALSE
 #' )
 #' check_values(dat, annots)